function_apps/ingest_sharepoint_files/ingest_sharepoint_files_blueprint.py
import logging
import os

import azure.functions as func

from models import Cards, Transactions, Users

ingest_sp_bp = func.Blueprint()


@ingest_sp_bp.function_name(name="IngestSharePointFilesTimer")
@ingest_sp_bp.schedule(
    schedule="0 */2 * * * *", arg_name="myTimer", run_on_startup=True, use_monitor=False
)
def timer_trigger(myTimer: func.TimerRequest) -> None:
    if myTimer.past_due:
        logging.info("The timer is past due!")

    logging.info("Simulate ingesting sharepoint file(s)")

    logging.info("Python timer trigger function executed.")
